fix: footer and storybook
