+++ packages/frontend/src/components/organisms/Footer.tsx
@@ -1,15 +1,4 @@
-import {
-  Box,
-  Button,
-  ButtonGroup,
-  chakra,
-  Divider,
-  SimpleGrid,
-  Stack,
-  Text,
-  useColorModeValue,
-  VisuallyHidden,
-} from "@chakra-ui/react";
+import { Box, ButtonGroup, chakra, Container, Stack, Text, useColorModeValue, VisuallyHidden } from "@chakra-ui/react";
 import { ReactNode } from "react";
 import React from "react";
 import { FaTwitter, FaMedium, FaDiscord, FaGithub } from "react-icons/fa";
@@ -39,65 +28,17 @@ const SocialButton = ({ children, label, href }: { children: ReactNode; label: s
 };
 
 export const Footer: React.FC = () => {
-  const links = [
-    {
-      title: "Product",
-      links: [
-        { label: "NFTHashi", href: "https://www.nfthashi.com/" },
-        // { label: "KaguraSwap", href: "#" },
-      ],
-    },
-    {
-      title: "Resources",
-      links: [
-        { label: "Docs", href: "https://docs.chocomint.app" },
-        { label: "Github", href: "https://github.com/chocomintapp" },
-        { label: "Gitcoin", href: "https://gitcoin.co/grants/4589/chocomint" },
-        { label: "DappRader", href: "https://dappradar.com/multichain/other/chocomint" },
-      ],
-    },
-    {
-      title: "Social",
-      links: [
-        { label: "Twitter", href: "https://twitter.com/chocomintapp" },
-        { label: "Discord", href: "https://discord.com/invite/UMcFBDfPa8" },
-        { label: "Medium", href: "https://medium.com/@Chocomintapp" },
-      ],
-    },
-  ];
-
   return (
-    <Box bg="teal.900" color="white" px={4}>
-      <Stack
-        justify="space-between"
-        align="start"
-        direction={{ base: "column", lg: "row" }}
-        py={{ base: "12", md: "16" }}
-        spacing="8"
+    <Box bg={useColorModeValue("gray.50", "gray.900")} color={useColorModeValue("gray.700", "gray.200")}>
+      <Container
+        as={Stack}
+        maxW={"6xl"}
+        py={4}
+        direction={{ base: "column", md: "row" }}
+        spacing={4}
+        justify={{ base: "center", md: "space-between" }}
+        align={{ base: "center", md: "center" }}
       >
-        <Stack spacing={{ base: "6", md: "8" }} align="start">
-          {/* <Logo /> */}
-          <Text color="on-accent-muted">Create NFT remarkably fast.</Text>
-        </Stack>
-        <SimpleGrid columns={{ base: 2, md: 3 }} gap="8" width={{ base: "full", lg: "auto" }}>
-          {links.map((group, idx) => (
-            <Stack key={idx} spacing="4" minW={{ lg: "40" }}>
-              <Text fontSize="sm" fontWeight="semibold" color="teal.50">
-                {group.title}
-              </Text>
-              <Stack spacing="3" shouldWrapChildren>
-                {group.links.map((link, idx) => (
-                  <Button key={idx} as="a" variant="gray.50" href={link.href} px="0" target="blank">
-                    {link.label}
-                  </Button>
-                ))}
-              </Stack>
-            </Stack>
-          ))}
-        </SimpleGrid>
-      </Stack>
-      <Divider borderColor="bg-accent-subtle" />
-      <Stack pt="8" pb="12" justify="space-between" direction={{ base: "column-reverse", md: "row" }} align="center">
         <Text fontSize="sm" color="on-accent-subtle">
           &copy; {new Date().getFullYear()} BlockBase,Inc. All rights reserved.
         </Text>
@@ -115,7 +56,7 @@ export const Footer: React.FC = () => {
             <FaMedium />
           </SocialButton>
         </ButtonGroup>
-      </Stack>
+      </Container>
     </Box>
   );
 };
