+++ packages/frontend/src/components/organisms/Resources.stories.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { MemoryRouter } from "react-router-dom";
+import { RecoilRoot } from "recoil";
+import { Resources } from "./Resources";
+
+export default {
+  title: "Organisms/Resources",
+  component: Resources,
+};
+
+export const Control: React.FC = (props) => (
+  <RecoilRoot>
+    <MemoryRouter>
+      <Resources {...props}>{props.children}</Resources>
+    </MemoryRouter>
+  </RecoilRoot>
+);
